+++ es6/tests/__snapshots.js
@@ -33,10 +33,10 @@ exports[`Api versioning should fail with invalid versions-1`] = {
     ],
     "currentModuleApiVersion": [
       3,
-      46,
+      47,
       0
     ],
-    "explanation": "moduleAPIVersionMismatch : needed=5.6.0, current=3.46.0"
+    "explanation": "moduleAPIVersionMismatch : needed=5.6.0, current=3.47.0"
   }
 }
 
@@ -48,15 +48,15 @@ exports[`Api versioning should fail with invalid versions-2`] = {
     "id": "api_version_error",
     "neededVersion": [
       3,
-      47,
+      48,
       0
     ],
     "currentModuleApiVersion": [
       3,
-      46,
+      47,
       0
     ],
-    "explanation": "moduleAPIVersionMismatch : needed=3.47.0, current=3.46.0"
+    "explanation": "moduleAPIVersionMismatch : needed=3.48.0, current=3.47.0"
   }
 }
 
@@ -68,15 +68,15 @@ exports[`Api versioning should fail with invalid versions-3`] = {
     "id": "api_version_error",
     "neededVersion": [
       3,
-      46,
+      47,
       100
     ],
     "currentModuleApiVersion": [
       3,
-      46,
+      47,
       0
     ],
-    "explanation": "moduleAPIVersionMismatch : needed=3.46.100, current=3.46.0"
+    "explanation": "moduleAPIVersionMismatch : needed=3.47.100, current=3.47.0"
   }
 }
 
@@ -1180,10 +1180,10 @@ exports[`Verify apiversion should fail with invalid api version`] = {
     ],
     "currentModuleApiVersion": [
       3,
-      46,
+      47,
       0
     ],
-    "explanation": "moduleAPIVersionMismatch : needed=3.92.0, current=3.46.0"
+    "explanation": "moduleAPIVersionMismatch : needed=3.92.0, current=3.47.0"
   }
 }